requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.11.12
flake8==7.3.0
frozenlist==1.8.0
greenlet==3.3.0
idna==3.11
librt==0.7.4
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
pathspec==0.12.1
propcache==0.4.1
psycopg==3.3.2
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
